remove redundant modifiers

+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/note/ZapPollNoteViewModel.kt
@@ -262,7 +262,7 @@ class PollNoteViewModel : ViewModel() {
             valueMinimum?.let { minimum ->
                 valueMaximum?.let { maximum ->
                     if (minimum != maximum) {
-                        options.add(((minimum + maximum) / 2).toLong())
+                        options.add(((minimum + maximum) / 2))
                     }
                 }
             }
